refactor: split manager orders router into read and write modules
